chore: 타이머 폴더구조 정리
